Refactor handler orchestration
